Run event loop outside async function
Running the event loop inside an async function blocks new futures from
being spawned in the examples using spawn_local. This change moves the
event loop out of the async function and runs it directly in main. To
still be able to initialize the wgpu-rs adapter and device asynchronously,
we therefore need to introduce a custom event that is sent to the event
loop once the wgpu-rs adapter and device have been created.

Fixes #454